Updating time function, getting time (SO depending) need to be changed. because it return an integer not a double which is needed in the algorithms

+++ ColorClusterSegmentationStereoTracking/CCSSTracking/CCSSTracking/timefuns/time.h
@@ -12,7 +12,7 @@
 
 namespace sysctrl
 {
-		typedef float TReal;
+		typedef double TReal;
 
         class STime
         {
@@ -42,10 +42,10 @@ namespace sysctrl
 
                 // Internal use.
 		#if defined(_linux) || defined (ANDROID)
-            int        mLastTime;
+            unsigned double mLastTime;
 		#endif
 		#if defined (WIN32)
-		    unsigned mLastTime;
+			unsigned double mLastTime;
 		#endif
         };
